Modularize js code
